backend/pyproject.toml
[project]
name = "cf-ai-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.51",
    "chonkie>=1.4.0",
    "dotenv>=0.9.9",
    "fastapi>=0.119.0",
    "finnhub-python>=2.4.25",
    "html2text>=2025.4.15",
    "langgraph>=0.6.10",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.0.18",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "timedelta>=2020.12.3",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
